feat: add useUrlState hook
